commit 7

+++ BE-LosQuebrachosApp/Entities/Chofer.cs
@@ -5,6 +5,7 @@
         public int Id { get; set; }
         public string Nombre { get; set; }
         public string Apellido { get; set; }
-        public int Cuit { get; set; }
+        public long Cuit { get; set; }
+        
     }
 }
